% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divergence.R
\name{get_divergence}
\alias{get_divergence}
\title{Function to Calculate the Kullback-Leibler Divergence of a Variable}
\usage{
get_divergence(p, q)
}
\arguments{
\item{p}{receives the observations of the reference variable p, which is a vector;}

\item{q}{receives the observations of the comparison variable q, which is a vector.}
}
\value{
Returns the Kullback-Leibler divergence as a numeric value
}
\description{
The Kullback-Leibler (KL) divergence is a statistical measure that quantifies the discrepancy between relative frequencies (or, more precisely, between two probability distributions).
}
\details{
Computes the Kullback-Leibler divergence as a numerical value.
}
\examples{
\dontrun{

# Kullback-Leibler Divergence

base::rm(list = ls())

p = c(0.14, 0.21, 0.14, 0.29, 0.19)
q = c(0.14, 0.20, 0.12, 0.33, 0.19)

divergence = kl::get_divergence(p, q)

divergence |> print()

}

}
