% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divergence_df.R
\name{get_divergence_df}
\alias{get_divergence_df}
\title{Function to Calculate the Kullback-Leibler Divergence of n Variables from a data.frame}
\usage{
get_divergence_df(p, q)
}
\arguments{
\item{p}{receives a set of reference variables in a data.frame (a KL divergence is calculated for each variable in the data.frame)}

\item{q}{receives a set of comparison variables in a data.frame (a KL divergence is calculated for each variable in the data.frame)}
}
\value{
Returns the Kullback-Leibler divergence as a numeric value
}
\description{
The Kullback-Leibler (KL) divergence is a statistical measure that quantifies the discrepancy between relative frequencies (or, more precisely, between two probability distributions).
}
\details{
Computes the Kullback-Leibler divergence as a numerical value.
}
\examples{
\dontrun{

# Kullback-Leibler Divergence

base::rm(list = ls())

p <- data.frame(var1 = c(25, 36, 45.6, 98.2, 12.8),
                var2 = c(25, 69, 52.1, 25, 14.5))

q <- data.frame(var1 = c(25, 36, 45.6, 98.2, 12.8),
                var2 = c(25, 69, 52, 26, 14))

divergence = kl::get_divergence_df(p, q)

divergence |> print()

}

}
