% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_divergence.R
\name{get_plot_divergence}
\alias{get_plot_divergence}
\title{Function for Comparative Plotting Between Probability Distributions}
\usage{
get_plot_divergence(p, q)
}
\arguments{
\item{p}{A data.frame containing the set of reference variables}

\item{q}{A data.frame containing the set of comparison variables}
}
\value{
A histogram plot displaying the distribution of values between variables in `p` and `q`.
}
\description{
Function for Comparative Plotting Between Probability Distributions
}
\examples{
\dontrun{


library(truncnorm)

n <- 1000

# Gerar os data frames p e q

p <- data.frame(var1 = rtruncnorm(n, mean = 0.0, sd = 1),
                var2 = rtruncnorm(n, mean = 0.1, sd = 2),
                var3 = rtruncnorm(n, mean = 0.5, sd = 3),
                var4 = rtruncnorm(n, mean = 0.9, sd = 4))

q <- data.frame(var1 = rtruncnorm(n, mean = 0.9, sd = 1),
                var2 = rtruncnorm(n, mean = 0.5, sd = 2),
                var3 = rtruncnorm(n, mean = 0.1, sd = 3),
                var4 = rtruncnorm(n, mean = 0.0, sd = 4))


 get_plot_divergence(p, q)

}

}
